backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
playwright>=1.36.0
firebase-admin>=6.2.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pydantic>=2.0.0